+++ l10n_br_hr/model/__init__.py
@@ -8,3 +8,4 @@ from . import hr_employee
 from . import hr_job
 from . import res_company
 from . import res_partner
+from . import hr_employee_dependent
