+++ l10n_br_hr/model/hr_employee.py
@@ -150,86 +150,3 @@ class HrEmployee(models.Model):
         default='funcionario',
         required=True,
     )
-
-
-class HrEmployeeDependent(models.Model):
-    _name = 'hr.employee.dependent'
-    _description = 'Employee\'s Dependents'
-    _rec_name = "name"
-    _inherits = {'res.partner': 'partner_id'}
-
-    # @api.constrains('dependent_cpf')
-    # def _validate_cpf(self):
-    #     if self.dependent_cpf:
-    #         if not fiscal.validate_cpf(self.dependent_cpf):
-    #             raise ValidationError(_('Invalid CPF for dependent %s'
-    #                                     % self.dependent_name))
-
-    partner_id = fields.Many2one(
-        string='Partner',
-        comodel_name='res.partner',
-        ondelete='cascade',
-        auto_join=True,
-        required=True,
-        help="Parceiro que contem as informações de bancárias do dependente."
-
-    )
-    employee_id = fields.Many2one(comodel_name='hr.employee',
-                                  string='Employee')
-    dependent_dob = fields.Date(string='Date of birth', required=True)
-    dependent_type_id = fields.Many2one(string='Relatedness',
-                                        required=True,
-                                        comodel_name='hr.dependent.type')
-    pension_benefits = fields.Float(string='Allowance value')
-    dependent_verification = fields.Boolean(string='IRPF (é dependente)')
-    health_verification = fields.Boolean(string='Healthcare plan')
-    dependent_gender = fields.Selection(string='Gender', selection=[
-        ('m', 'Male'),
-        ('f', 'Female')])
-
-    have_alimony = fields.Boolean(string='Tem Pensão?')
-
-    partner_id_bank_ids = fields.One2many(
-        comodel_name='res.partner.bank',
-        string='Info Bank',
-        related='partner_id.bank_ids',
-    )
-    dep_sf = fields.Boolean(
-        string='Salário Família?',
-    )
-    inc_trab = fields.Boolean(
-        string='Incapacidade Física ou Mental?',
-    )
-    inc_trab_inss_file = fields.Binary(
-        string='Atestado de incapacidade INSS',
-    )
-    precisa_cpf = fields.Boolean(
-        string='Precisa passar cpf?',
-        compute='_precisa_preencher_cpf'
-    )
-    relative_file = fields.Binary(
-        string='Documento comprobatório da relação',
-        help='Certidão de Nascimento / Casamento / etc'
-    )
-
-    @api.multi
-    @api.depends('dependent_dob', 'dependent_verification')
-    def _precisa_preencher_cpf(self):
-        for record in self:
-            precisa_cpf = False
-            if record.dependent_verification and record.dependent_dob:
-                data_atual = fields.Datetime.from_string(fields.Datetime.now())
-                data_aniversario = fields.Datetime.from_string(
-                    record.dependent_dob)
-                if (data_atual - data_aniversario).days / 365 >= 8:
-                    precisa_cpf = True
-            record.precisa_cpf = precisa_cpf
-
-    @api.model
-    def create(self, vals):
-        ctx = self.env.context.copy()
-        ctx['create_depentent'] = True
-        ctx['depentent_employee_id'] = vals.get('employee_id', False)
-        patient = super(
-            HrEmployeeDependent, self.with_context(ctx)).create(vals)
-        return patient
