+++ l10n_br_hr/model/res_partner.py
@@ -2,7 +2,7 @@
 # (c) 2019 KMEE INFORMATICA LTDA
 # License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
 
-from openerp import models, fields, api, _
+from openerp import models, fields, api
 
 
 class ResPartner(models.Model):
